Moving status mapper to event subscriber
